storage: move Error enum from page_id.rs into its own error.rs
Error lived alongside PageId in page_id.rs for no good reason — they are
independent types. Separating them makes each file's purpose clear.
